market-cap-ppt-generator/requirements.txt
streamlit
python-pptx
pandas
openpyxl